+++ changelog.md
@@ -33,6 +33,7 @@
 * LLT-4858: NAT state analytics reporting
 * LLT-5038: Add doc rs as in-repo documentation system skeleton
 * LLT-5034: Upgrade rust toolchain to v1.77.2
+* LLT-4924: Enable size reduction compiler flags
 
 <br>